Remove ElementDetailsActivity
