+++ app/src/main/res/layout-sw600dp/activity_home.xml
@@ -14,7 +14,7 @@
 
     <fragment
         android:id="@+id/description_fragment"
-        android:name="com.example.davit_zakaryan.mvvmapp.ui.elements.ElementsFragment"
+        android:name="com.example.davit_zakaryan.mvvmapp.ui.element_details.ElementDetailsFragment"
         android:layout_width="0dp"
         android:layout_height="match_parent"
         android:layout_weight="2" />
