+++ app/src/main/res/layout/activity_details.xml
@@ -1,96 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:fresco="http://schemas.android.com/apk/res-auto">
-
-    <data>
-
-        <variable
-            name="element"
-            type="com.example.davit_zakaryan.mvvmapp.data.model.Element" />
-    </data>
-
-    <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <include
-            layout="@layout/layout_toolbar"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <com.facebook.drawee.view.SimpleDraweeView
-            android:id="@+id/imageView"
-            android:layout_width="match_parent"
-            android:layout_height="190dp"
-            android:layout_marginTop="?attr/actionBarSize"
-            android:contentDescription="@null"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            fresco:actualImageScaleType="focusCrop"
-            fresco:imageUrl="@{element.url}"
-            fresco:placeholderImage="@drawable/ic_launcher_background"
-            fresco:placeholderImageScaleType="centerCrop"
-            tools:background="@color/colorAccent" />
-
-        <TextView
-            android:id="@+id/activity_details_short_desc_text"
-            android:layout_width="195dp"
-            android:layout_height="40dp"
-            android:layout_marginBottom="8dp"
-            android:layout_marginStart="24dp"
-            android:text="@{element.shortDesc}"
-            android:textColor="@android:color/white"
-            app:layout_constraintBottom_toBottomOf="@+id/imageView"
-            app:layout_constraintStart_toStartOf="parent"
-            tools:text="Short description" />
-
-        <TextView
-            android:id="@+id/activity_details_name_text"
-            android:layout_width="170dp"
-            android:layout_height="34dp"
-            android:layout_marginBottom="8dp"
-            android:layout_marginStart="24dp"
-            android:text="@{element.name}"
-            android:textColor="@android:color/white"
-            app:layout_constraintBottom_toTopOf="@+id/activity_details_short_desc_text"
-            app:layout_constraintStart_toStartOf="parent"
-            tools:text="Name" />
-
-        <TextView
-            android:id="@+id/activity_details_description_text"
-            android:layout_width="192dp"
-            android:layout_height="66dp"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="8dp"
-            android:text="@{element.description}"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/imageView"
-            tools:text="Long description" />
-
-        <com.example.custom_view.GaugeView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="8dp"
-            app:gv_level="@{element.level}"
-            app:gv_scalesCount="10"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/activity_details_description_text" />
-
-        <android.support.design.widget.FloatingActionButton
-            android:id="@+id/fab"
-            android:layout_width="wrap_content"
-            android:layout_height="64dp"
-            android:layout_gravity="bottom|end"
-            android:layout_margin="@dimen/fab_margin"
-            android:layout_marginBottom="8dp"
-            android:layout_marginEnd="8dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:srcCompat="@drawable/ic_edit_white_24px" />
-
-    </android.support.constraint.ConstraintLayout>
-</layout>
+<fragment xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/description_fragment"
+    android:name="com.example.davit_zakaryan.mvvmapp.ui.element_details.ElementDetailsFragment"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent" />
