+++ app/src/main/res/layout/fragment_elements.xml
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context="com.example.davit_zakaryan.mvvmapp.ui.elements.ElementsActivity">
+        tools:context="com.example.davit_zakaryan.mvvmapp.ui.base.HomeActivity">
 
         <include
             android:id="@+id/toolbar"
